+++ docs_src/src/guide/changelog.md
@@ -61,6 +61,7 @@ meta:
 - Re-organised the `goyave.Response` structure fields to save some memory.
 - Added the ability to regsiter new SQL dialects to use with GORM.
 - `database.Close()` can now return errors.
+- Added `response.GetStacktrace()`, `response.IsEmpty()` and `response.IsHeaderWritten()`.
 
 ## v2.10.x